Title: 세 수의 합과 평균 구하기, Time: 127ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240717/세 수의 합과 평균 구하기/sum-and-mean-of-three-numbers.java
@@ -8,7 +8,7 @@ public class Main {
     int c=sc.nextInt();
 
     int sum;
-    double avg;
+    int avg;
     sum=a+b+c;
     avg=sum/3;
     System.out.println(sum+"\n"+avg);
